refactor: put sample fixture in tests/__init__.py for organization
