+++ tests/__init__.py
@@ -57,4 +57,20 @@ def thing():
         session.close()
 
 
+@pytest.fixture
+def sample_fixture(thing):
+    with session_ctx() as session:
+        sample = Sample(
+            thing_id=thing.id,
+            collection_timestamp="2025-01-01T00:00:00+00:00",
+            collection_method="manual",
+        )
+        session.add(sample)
+        session.commit()
+        session.refresh(sample)
+        yield thing, sample
+        session.delete(sample)
+        session.commit()
+
+
 # ============= EOF =============================================
